requirements.txt
biopython==1.78
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
et-xmlfile==1.0.1
firebase-admin==4.5.3
google-api-core==1.26.3
google-api-python-client==2.2.0
google-auth==1.29.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.0
google-cloud-storage==1.37.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.37.0
httplib2==0.19.1
idna==2.10
jdcal==1.4.1
msgpack==1.0.2
numpy==1.20.2
openpyxl==3.0.7 
packaging==20.9
proto-plus==1.18.1
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-Levenshtein==0.12.2
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
textdistance==4.2.1
uritemplate==3.0.1
urllib3==1.26.4
wincertstore==0.2
